Update readme-qt.rst
In documentation added the option "USE_UPNP=-" which disables UPNP. Otherwise the compilation exits with the error because miniupnpc is required but not installed. Tested following the readme instructions from clean kubuntu14.04 install. 
I know that UPNP support is recommended but this is explained further down in documentation so the user can enable it if he wishes. It is more important for new users to be able to compile without errors out of the box following the instructions.